Simplify Darwin code dummy/fakes
