--- pkg/ebpf/instrumenter_linux.go
@@ -1,44 +1,0 @@
-package ebpf
-
-import (
-	"encoding/binary"
-	"syscall"
-	"unsafe"
-
-	"github.com/cilium/ebpf"
-	"github.com/grafana/ebpf-autoinstrument/pkg/exec"
-	"github.com/prometheus/procfs"
-	"golang.org/x/sys/unix"
-)
-
-func attachSocketFilter(filter *ebpf.Program) (int, error) {
-	fd, err := unix.Socket(unix.AF_PACKET, unix.SOCK_RAW, int(htons(unix.ETH_P_ALL)))
-	if err == nil {
-		ssoErr := syscall.SetsockoptInt(fd, unix.SOL_SOCKET, unix.SO_ATTACH_BPF, filter.FD())
-		if ssoErr != nil {
-			return -1, ssoErr
-		}
-		return fd, nil
-	}
-
-	return -1, err
-}
-
-func isLittleEndian() bool {
-	var a uint16 = 1
-
-	return *(*byte)(unsafe.Pointer(&a)) == 1
-}
-
-func htons(a uint16) uint16 {
-	if isLittleEndian() {
-		var arr [2]byte
-		binary.LittleEndian.PutUint16(arr[:], a)
-		return binary.BigEndian.Uint16(arr[:])
-	}
-	return a
-}
-
-func processMaps(pid int32) ([]*procfs.ProcMap, error) {
-	return exec.FindLibMaps(pid)
-}
